backend/requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
Django==3.1.5
django-cleanup==5.1.0
django-cors-headers==3.6.0
django-rest-framework==0.1.0
djangorestframework==3.12.2
docker==4.4.1
idna==2.10
numpy==1.19.5
pip==20.3.3
PyMySQL==1.0.2
pytz==2020.5
requests==2.25.1
setuptools==49.2.1
six==1.15.0
SQLAlchemy==1.3.22
sqlparse==0.4.1
urllib3==1.26.2
websocket-client==0.57.0
cryptography
uWSGI

